feat(consensus): add a channel to transfer events to reactor
