+++ src/noir/consensus/types.h
@@ -89,166 +89,6 @@ struct round_vote_set {
   std::shared_ptr<vote_set> precommits;
 };
 
-/**
- * Keeps track of all VoteSets from round 0 to round 'round'.
- * Also keeps track of up to one RoundVoteSet greater than
- * 'round' from each peer, to facilitate catchup syncing of commits.
- *
- * A commit is +2/3 precommits for a block at a round,
- * but which round is not known in advance, so when a peer
- * provides a precommit for a round greater than mtx.round,
- * we create a new entry in roundVoteSets but also remember the
- * peer to prevent abuse.
- * We let each peer provide us with up to 2 unexpected "catchup" rounds.
- * One for their LastCommit round, and another for the official commit round.
- */
-struct height_vote_set {
-  std::string chain_id;
-  int64_t height;
-  validator_set val_set;
-
-  std::mutex mtx;
-  int32_t round;
-  std::map<int32_t, round_vote_set> round_vote_sets;
-  std::map<node_id, std::vector<int32_t>> peer_catchup_rounds;
-
-  static std::shared_ptr<height_vote_set> new_height_vote_set(
-    std::string chain_id_, int64_t height_, const validator_set& val_set_) {
-    auto hvs = std::make_shared<height_vote_set>();
-    hvs->chain_id = chain_id_;
-    hvs->reset(height_, val_set_);
-    return hvs;
-  }
-
-  void reset(int64_t height_, const validator_set& val_set_) {
-    std::lock_guard<std::mutex> g(mtx);
-    height = height_;
-    val_set = val_set_;
-    add_round(0);
-    round = 0;
-  }
-
-  void add_round(int32_t round_) {
-    if (round_vote_sets.contains(round_))
-      throw std::runtime_error("add_round() for an existing round");
-
-    auto prevotes = vote_set::new_vote_set(chain_id, height, round_, p2p::Prevote, val_set);
-    auto precommits = vote_set::new_vote_set(chain_id, height, round_, p2p::Precommit, val_set);
-    round_vote_sets[round_] = round_vote_set{prevotes, precommits};
-  }
-
-  void set_round(int32_t round_) {
-    std::lock_guard<std::mutex> g(mtx);
-    auto new_round_ = round - 1; // todo - safe subtract
-    if (round != 0 && round_ < new_round_)
-      throw std::runtime_error("set_round() must increment round");
-    for (auto r = new_round_; r <= round_; r++) {
-      if (round_vote_sets.contains(r))
-        continue; // Already exists because peer_catchup_rounds
-      add_round(r);
-    }
-    round = round_;
-  }
-
-  std::shared_ptr<vote_set> get_vote_set(int32_t round_, p2p::signed_msg_type vote_type) {
-    auto it = round_vote_sets.find(round_);
-    if (it == round_vote_sets.end())
-      return nullptr;
-    switch (vote_type) {
-    case p2p::Prevote:
-      return it->second.prevotes;
-    case p2p::Precommit:
-      return it->second.precommits;
-    default:
-      throw std::runtime_error(fmt::format("get_vote_set() unexpected vote type {}", vote_type));
-    }
-  }
-
-  /**
-   * return last round number or -1 if not exists
-   */
-  int32_t pol_info() {
-    std::lock_guard<std::mutex> g(mtx);
-    for (auto r = round; r >= 0; r--) {
-      auto rvs = get_vote_set(r, p2p::Prevote);
-      if (rvs->two_thirds_majority().has_value())
-        return r;
-    }
-    return -1;
-  }
-
-  bool add_vote(vote vote_, node_id peer_id) {
-    std::lock_guard<std::mutex> g(mtx);
-    if (!p2p::is_vote_type_valid(vote_.type))
-      return false;
-    auto vote_set_ = get_vote_set(vote_.round, vote_.type);
-    if (vote_set_ == nullptr) {
-      auto it = peer_catchup_rounds.find(peer_id);
-      if (it == peer_catchup_rounds.end() || it->second.size() >= 2) {
-        // punish peer // todo - how?
-        elog("peer has sent a vote that does not match our round for more than one round");
-        return false;
-      }
-      add_round(vote_.round);
-      vote_set_ = get_vote_set(vote_.round, vote_.type);
-      it->second.push_back(vote_.round);
-    }
-    return vote_set_->add_vote(vote_);
-  }
-
-  std::shared_ptr<vote_set> prevotes(int32_t round_) {
-    std::lock_guard<std::mutex> g(mtx);
-    return get_vote_set(round_, p2p::Prevote);
-  }
-
-  std::shared_ptr<vote_set> precommits(int32_t round_) {
-    std::lock_guard<std::mutex> g(mtx);
-    return get_vote_set(round_, p2p::Precommit);
-  }
-};
-
-/**
- * Defines the internal consensus state.
- * NOTE: not thread safe
- */
-struct round_state {
-  int64_t height;
-  int32_t round;
-  p2p::round_step_type step;
-  p2p::tstamp start_time;
-
-  // Subjective time when +2/3 precommits for Block at Round were found
-  p2p::tstamp commit_time;
-  std::shared_ptr<validator_set> validators{};
-  std::shared_ptr<p2p::proposal_message> proposal{};
-  std::shared_ptr<block> proposal_block{};
-  std::shared_ptr<part_set> proposal_block_parts{};
-  int32_t locked_round;
-  std::shared_ptr<block> locked_block{};
-  std::shared_ptr<part_set> locked_block_parts{};
-
-  // Last known round with POL for non-nil valid block.
-  int32_t valid_round;
-  std::shared_ptr<block> valid_block{}; // Last known block of POL mentioned above.
-
-  std::shared_ptr<part_set> valid_block_parts{};
-  std::shared_ptr<height_vote_set> votes{};
-  int32_t commit_round;
-  std::shared_ptr<vote_set> last_commit{};
-  std::shared_ptr<validator_set> last_validators{};
-  bool triggered_timeout_precommit;
-
-  struct event_data {
-    int64_t height;
-    int32_t round;
-    int32_t step;
-  };
-
-  event_data new_event_data() {
-    return {.height = height, .round = round, .step = static_cast<int32_t>(step)};
-  }
-};
-
 inline p2p::tstamp get_time() {
   return std::chrono::system_clock::now().time_since_epoch().count();
 }
@@ -265,51 +105,6 @@ struct proposal : p2p::proposal_message {
 
   static proposal new_proposal(int64_t height_, int32_t round_, int32_t pol_round_, p2p::block_id b_id_) {
     return proposal{p2p::Proposal, height_, round_, pol_round_, b_id_, get_time()};
-  }
-};
-
-/// \brief contains known state of a peer
-struct peer_round_state {
-  int64_t height{};
-  int32_t round{};
-  p2p::round_step_type step{};
-
-  p2p::tstamp start_time{};
-
-  bool proposal{};
-  p2p::part_set_header proposal_block_part_set_header;
-  std::shared_ptr<bit_array> proposal_block_parts;
-  int32_t proposal_pol_round{};
-
-  std::shared_ptr<bit_array> proposal_pol;
-  std::shared_ptr<bit_array> prevotes;
-  std::shared_ptr<bit_array> precommits;
-  int32_t last_commit_round{};
-  std::shared_ptr<bit_array> last_commit;
-
-  int32_t catchup_commit_round{};
-
-  std::shared_ptr<bit_array> catchup_commit;
-
-  peer_round_state() = default;
-  peer_round_state(const peer_round_state&) = default;
-  peer_round_state& operator=(const peer_round_state& v) {
-    proposal_block_part_set_header = v.proposal_block_part_set_header; // TODO: check if this is correct
-
-    height = v.height;
-    round = v.round;
-    step = v.step;
-    start_time = v.start_time;
-    proposal = v.proposal;
-    proposal_pol_round = v.proposal_pol_round;
-    last_commit_round = v.last_commit_round;
-    catchup_commit_round = v.catchup_commit_round;
-    proposal_block_parts = v.proposal_block_parts->copy();
-    proposal_pol = v.proposal_pol->copy();
-    prevotes = v.prevotes->copy();
-    precommits = v.precommits->copy();
-    last_commit = v.last_commit->copy();
-    catchup_commit = v.catchup_commit->copy();
   }
 };
